Split tests and limit tests that run in CI
